+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 - Added support for processing patterns with big stacksizes
 - The slot where the Wireless Grid is in in the Wireless Grid GUI is now disabled, so the item can't be thrown out of the inventory by accident
 - Changed Relay recipe to use redstone torch instead of Basic Processor
+- Placed machines now face the block they are placed on, like hoppers
 
 **NOTE:** Config change: the config options for the energy capacity of the Controller and wether the Controller uses energy are now in a different config category called "controller", if you changed these config options, don't forget the change it under the new category.